Extract answers/questions/submissions/suggest-edits into services
Thin endpoints for answers, questions, submissions, answer_suggest_edits,
and me (prior). Fix bot submission create via services. Fix shaper
principal_view after materializer alias removal.
