+++ chafan_core/app/api/api_v1/endpoints/answers.py
@@ -1,26 +1,21 @@
-from typing import Any, List, Union
+from typing import Any, List, Optional
 
-from fastapi import APIRouter, Depends, Query, Request, Response, BackgroundTasks
-from fastapi.encoders import jsonable_encoder
+from fastapi import APIRouter, BackgroundTasks, Depends, Query, Request, Response
 from sqlalchemy.orm import Session
 
-from chafan_core.app import crud, models, schemas, view_counters
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
+from chafan_core.app.common import client_ip
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.services import answers as answers_service
-from chafan_core.app.common import OperationType, client_ip
-from chafan_core.app.endpoint_utils import check_writing_session
 from chafan_core.app.limiter import limiter
-from chafan_core.app.responders.archives import answer_archive_schema_from_orm
-from chafan_core.app.user_permission import check_user_in_site
 from chafan_core.app.schemas.answer import AnswerModUpdate
-from chafan_core.app.schemas.event import EventInternal, UpvoteAnswerInternal
-from chafan_core.app.schemas.richtext import RichText
-from chafan_core.utils.base import HTTPException_, filter_not_none, get_utc_now, unwrap
-from chafan_core.utils.constants import MAX_ARCHIVE_PAGINATION_LIMIT
+from chafan_core.app.services import answers as answers_service
 from chafan_core.app.services.postprocess import postprocess_new_answer
+from chafan_core.utils.base import HTTPException_
+from chafan_core.utils.constants import MAX_ARCHIVE_PAGINATION_LIMIT
 
 import logging
+
 logger = logging.getLogger(__name__)
 
 router = APIRouter()
@@ -35,8 +30,6 @@ def get_one(
     """
     Get answer in one of current_user's belonging sites.
     """
-    from chafan_core.app.services import answers as answers_service
-
     answer_data = answers_service.get_answer_schema(ctx, uuid)
     if answer_data is None:
         raise HTTPException_(
@@ -52,8 +45,6 @@ def delete_answer(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     uuid: str,
 ) -> Any:
-    from chafan_core.app.services import answers as answers_service
-
     error_msg = answers_service.delete_answer(
         ctx.get_db(),
         uuid=uuid,
@@ -77,23 +68,8 @@ def get_answer_draft(
     """
     Get answer's draft body as its author.
     """
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    if current_user_id != answer.author_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    draft = None
-    if answer.body_draft:
-        draft = RichText(source=answer.body_draft, editor=answer.draft_editor)
-    return schemas.answer.AnswerDraft(
-        draft_saved_at=answer.draft_saved_at,
-        content_draft=draft,
+    return answers_service.get_draft(
+        db, uuid=uuid, principal_id=current_user_id
     )
 
 
@@ -104,34 +80,9 @@ def delete_answer_draft(
     uuid: str,
     current_user_id: int = Depends(deps.get_current_user_id),
 ) -> Any:
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    if current_user_id != answer.author_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    if not answer.body_draft:
-        raise HTTPException_(
-            status_code=400,
-            detail="Answer has no draft.",
-        )
-    data = schemas.answer.AnswerDraft(
-        draft_saved_at=answer.draft_saved_at,
-        content_draft=RichText(
-            source=answer.body_draft,
-            editor=answer.draft_editor,
-        ),
+    return answers_service.delete_draft(
+        db, uuid=uuid, principal_id=current_user_id
     )
-    answer.body_draft = None
-    answer.draft_saved_at = None
-    db.add(answer)
-    db.commit()
-    return data
 
 
 @router.get("/{uuid}/archives/", response_model=List[schemas.AnswerArchive])
@@ -148,21 +99,9 @@ def get_answer_archives(
     """
     Get answer's archives as its author.
     """
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    if current_user_id != answer.author_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    return [
-        answer_archive_schema_from_orm(a)
-        for a in answer.archives[skip : (skip + limit)]
-    ]
+    return answers_service.list_archives(
+        db, uuid=uuid, principal_id=current_user_id, skip=skip, limit=limit
+    )
 
 
 @router.post("/{uuid}/views/", response_model=schemas.GenericResponse)
@@ -174,13 +113,7 @@ def bump_views_counter(
     uuid: str,
     ctx: RequestContext = Depends(deps.get_request_context),
 ) -> Any:
-    answer = crud.answer.get_by_uuid(ctx.get_db(), uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    view_counters.add_view_async(ctx, "answer", answer.id)
+    answers_service.bump_views(ctx, uuid=uuid)
     return schemas.GenericResponse()
 
 
@@ -195,99 +128,13 @@ def create_answer(
     """
     Create new answer authored by the current user in one of the belonging sites.
     """
-    current_user_id = ctx.unwrapped_principal_id()
-    crud.audit_log.create_with_user(
-        ctx.get_db(),
-        ipaddr=client_ip(request),
-        user_id=current_user_id,
-        api="post answer",
-        request_info={"answer_in": jsonable_encoder(answer_in)},
+    answer, data, needs_postprocess = answers_service.create_answer(
+        ctx, answer_in=answer_in, ipaddr=client_ip(request)
     )
-
-    question = crud.question.get_by_uuid(
-        ctx.get_db(), uuid=answer_in.question_uuid
-    )
-    if not question:
-        raise HTTPException_(
-            status_code=400,
-            detail="The question doesn't exist in the system.",
-        )
-    check_writing_session(answer_in.writing_session_uuid)
-    check_user_in_site(
-        ctx.get_db(),
-        site=question.site,
-        user_id=current_user_id,
-        op_type=OperationType.WriteSiteAnswer,
-    )
-    if any(
-        answer.author_id == current_user_id
-        for answer in question.answers
-        if not answer.is_deleted
-    ):
-        raise HTTPException_(
-            status_code=400,
-            detail="You have saved an answer before.",
-        )
-    answer = crud.answer.create_with_author(
-        ctx.get_db(),
-        obj_in=answer_in,
-        author_id=current_user_id,
-        site_id=question.site_id,
-    )
-    if answer.is_published:
+    if needs_postprocess:
         logger.info(f"create_answer add postprocess task id={answer.id}")
         background_tasks.add_task(postprocess_new_answer, answer.id, False)
-    return answers_service.answer_schema(ctx, answer)
-
-
-def _update_answer(
-    ctx: RequestContext,
-    *,
-    answer: models.Answer,
-    answer_in: schemas.AnswerUpdate,
-    background_tasks: BackgroundTasks,
-) -> schemas.Answer:
-    db = ctx.get_db()
-    answer_in_dict = answer_in.dict(exclude_none=True)
-    if answer_in.is_draft and answer_in.updated_content:
-        del answer_in_dict["updated_content"]
-        answer_in_dict["body_draft"] = answer_in.updated_content.source
-        answer_in_dict["draft_editor"] = answer_in.updated_content.editor
-        answer_in_dict["draft_saved_at"] = get_utc_now()
-    else:
-        if answer.is_published:
-            archive = models.Archive(
-                editor=answer.editor,
-                answer_id=answer.id,
-                body=answer.body,
-                created_at=answer.updated_at,
-            )
-            db.add(archive)
-            answer.archives.append(archive)
-            db.commit()
-        answer_in_dict["is_published"] = True
-        answer_in_dict["updated_at"] = get_utc_now()
-
-        if answer_in.updated_content:
-            del answer_in_dict["updated_content"]
-            answer_in_dict["body"] = answer_in.updated_content.source
-            answer_in_dict[
-                "body_prerendered_text"
-            ] = answer_in.updated_content.rendered_text
-            answer_in_dict["editor"] = answer_in.updated_content.editor
-
-        answer_in_dict["body_draft"] = None
-        answer_in_dict["draft_saved_at"] = None
-
-    was_published = answer.is_published
-    answer = crud.answer.update_checked(db, db_obj=answer, obj_in=answer_in_dict)
-
-    if answer.is_published:
-        # NOTE: Since is_published will not be reverted, thus this should only be delivered once
-        # TODO: Implement the update subscription logic
-
-        background_tasks.add_task(postprocess_new_answer, answer.id, was_published)
-    return unwrap(answers_service.answer_schema(ctx, answer))
+    return data
 
 
 @router.put("/{uuid}", response_model=schemas.Answer)
@@ -302,39 +149,12 @@ def update_answer(
     """
     Update answer authored by current user in one of current user's belonging sites.
     """
-    db = ctx.get_db()
-    current_user_id = ctx.unwrapped_principal_id()
-    crud.audit_log.create_with_user(
-        db,
-        ipaddr=client_ip(request),
-        user_id=current_user_id,
-        api="post answer",
-        request_info={"answer_in": jsonable_encoder(answer_in), "uuid": uuid},
+    answer, data, needs_postprocess, was_published = answers_service.update_answer(
+        ctx, uuid=uuid, answer_in=answer_in, ipaddr=client_ip(request)
     )
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    if answer.author_id != current_user_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer is not authored by current user.",
-        )
-    question = crud.question.get_by_id(db, id=answer.question_id)
-    if not question:
-        raise HTTPException_(
-            status_code=400,
-            detail="The question doesn't exist in the system.",
-        )
-    check_user_in_site(
-        db,
-        site=question.site,
-        user_id=current_user_id,
-        op_type=OperationType.WriteSiteAnswer,
-    )
-    return _update_answer(ctx, answer=answer, answer_in=answer_in, background_tasks=background_tasks)
+    if needs_postprocess:
+        background_tasks.add_task(postprocess_new_answer, answer.id, was_published)
+    return data
 
 
 @router.put("/{uuid}/mod", response_model=schemas.Answer, include_in_schema=False)
@@ -347,34 +167,9 @@ def update_answer_by_mod(
     """
     Update answer as moderator of the site.
     """
-    db = ctx.get_db()
-    current_user_id = ctx.principal_id
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    answer_data = answers_service.answer_schema(ctx, answer)
-    if answer_data is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    site = crud.site.get_by_id(db, id=answer.site_id)
-    if not site:
-        # The site doesn't exist in the system.
-        return False
-    if site.moderator_id != current_user_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    answer = crud.answer.update_checked(
-        db, db_obj=answer, obj_in=update_in.dict(exclude_none=True)
+    return answers_service.update_answer_by_mod(
+        ctx, uuid=uuid, update_in=update_in
     )
-    answer_data = answers_service.answer_schema(ctx, answer)
-    return answer_data
 
 
 @router.get("/{uuid}/upvotes/", response_model=schemas.AnswerUpvotes)
@@ -383,8 +178,6 @@ def get_answer_upvotes(
     ctx: RequestContext = Depends(deps.get_request_context),
     uuid: str,
 ) -> Any:
-    from chafan_core.app.services import answers as answers_service
-
     data = answers_service.get_answer_upvotes(
         ctx.get_db(),
         uuid=uuid,
@@ -407,89 +200,7 @@ def upvote_answer(
     """
     Upvote answer as the current user in one of current user's belonging sites.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    upvoted = (
-        db.query(models.Answer_Upvotes)
-        .filter_by(answer_id=answer.id, voter_id=current_user.id, cancelled=False)
-        .first()
-        is not None
-    )
-    if not upvoted:
-        if current_user.id == answer.author_id:
-            raise HTTPException_(
-                status_code=400,
-                detail="Author can't upvote authored answer.",
-            )
-        if current_user.remaining_coins < answer.site.upvote_answer_coin_deduction:
-            raise HTTPException_(
-                status_code=400,
-                detail="Insufficient coins.",
-            )
-        question = crud.question.get_by_id(db, id=answer.question_id)
-        if not question:
-            raise HTTPException_(
-                status_code=400,
-                detail="The question doesn't exist in the system.",
-            )
-        check_user_in_site(
-            db,
-            site=question.site,
-            user_id=current_user.id,
-            op_type=OperationType.ReadSite,
-        )
-        upvoted_before = (
-            db.query(models.Answer_Upvotes)
-            .filter_by(answer_id=answer.id, voter_id=current_user.id)
-            .first()
-            is not None
-        )
-        # Don't swap the statements before and after!
-        answer = crud.answer.upvote(db, db_obj=answer, voter=current_user)
-        if not upvoted_before:
-            crud.coin_payment.make_payment(
-                db,
-                obj_in=schemas.CoinPaymentCreate(
-                    payee_id=answer.author_id,
-                    amount=answer.site.upvote_answer_coin_deduction,
-                    event_json=EventInternal(
-                        created_at=get_utc_now(),
-                        content=UpvoteAnswerInternal(
-                            subject_id=current_user.id,
-                            answer_id=answer.id,
-                        ),
-                    ).json(),
-                ),
-                payer=current_user,
-                payee=answer.author,
-            )
-            crud.notification.create_with_content(
-                ctx,
-                receiver_id=answer.author.id,
-                event=EventInternal(
-                    created_at=get_utc_now(),
-                    content=UpvoteAnswerInternal(
-                        subject_id=current_user.id,
-                        answer_id=answer.id,
-                    ),
-                ),
-            )
-        db.commit()
-        db.refresh(answer)
-    valid_upvotes = (
-        db.query(models.Answer_Upvotes)
-        .filter_by(answer_id=answer.id, cancelled=False)
-        .count()
-    )
-    return schemas.AnswerUpvotes(
-        answer_uuid=answer.uuid, count=valid_upvotes, upvoted=True
-    )
+    return answers_service.upvote_answer(ctx, uuid=uuid)
 
 
 @router.delete("/{uuid}/upvotes/", response_model=schemas.AnswerUpvotes)
@@ -501,44 +212,7 @@ def cancel_upvote_answer(
     """
     Cancel upvote for answer as the current user in one of current user's belonging sites.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    answer = crud.answer.get_by_uuid(db, uuid=uuid)
-    if answer is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The answer doesn't exist in the system.",
-        )
-    upvoted = (
-        db.query(models.Answer_Upvotes)
-        .filter_by(answer_id=answer.id, voter_id=current_user.id, cancelled=False)
-        .first()
-        is not None
-    )
-    if upvoted:
-        question = crud.question.get_by_id(db, id=answer.question_id)
-        if not question:
-            raise HTTPException_(
-                status_code=400,
-                detail="The question doesn't exist in the system.",
-            )
-        check_user_in_site(
-            db,
-            site=question.site,
-            user_id=current_user.id,
-            op_type=OperationType.ReadSite,
-        )
-        answer = crud.answer.cancel_upvote(db, db_obj=answer, voter=current_user)
-        db.commit()
-        db.refresh(answer)
-    valid_upvotes = (
-        db.query(models.Answer_Upvotes)
-        .filter_by(answer_id=answer.id, cancelled=False)
-        .count()
-    )
-    return schemas.AnswerUpvotes(
-        answer_uuid=answer.uuid, count=valid_upvotes, upvoted=False
-    )
+    return answers_service.cancel_upvote_answer(ctx, uuid=uuid)
 
 
 @router.get("/{uuid}/suggestions/", response_model=List[schemas.AnswerSuggestEdit])
@@ -547,6 +221,4 @@ def get_suggestions(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     uuid: str,
 ) -> Any:
-    from chafan_core.app.services import answers as answers_service
-
     return answers_service.list_suggest_edits(ctx, uuid=uuid)
