+++ chafan_core/app/api/api_v1/endpoints/questions.py
@@ -1,51 +1,19 @@
-import datetime
-from typing import Any, List, Optional, Union
+from typing import Any, List, Optional
 
-from fastapi import APIRouter, Depends, Request, Response, BackgroundTasks
-from fastapi.encoders import jsonable_encoder
+from fastapi import APIRouter, BackgroundTasks, Depends, Request, Response
 
-from chafan_core.app import crud, models, schemas, view_counters
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
-from chafan_core.app.services import answers as answers_service
+from chafan_core.app.common import client_ip
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.common import OperationType, client_ip
-from chafan_core.app.endpoint_utils import get_site
 from chafan_core.app.limiter import limiter
-from chafan_core.app.user_permission import check_user_in_site, user_in_site
-from chafan_core.app.recs.ranking import rank_answers
-from chafan_core.app.schemas.event import (
-    EventInternal,
-    InviteAnswerInternal,
-    UpvoteQuestionInternal,
-)
 from chafan_core.app.services import questions as questions_service
-from chafan_core.app.services.postprocess import postprocess_new_question, postprocess_updated_question
-from chafan_core.utils.base import HTTPException_, filter_not_none
-
-import logging
-logger = logging.getLogger(__name__)
+from chafan_core.app.services.postprocess import (
+    postprocess_new_question,
+    postprocess_updated_question,
+)
 
 router = APIRouter()
-
-AnswersData = List[schemas.AnswerPreview]
-
-
-def _get_answers(
-    ctx: RequestContext, question: models.Question
-) -> List[schemas.AnswerPreview]:
-    return questions_service.list_answer_previews(ctx, question)
-
-
-def _get_question_data(
-    ctx: RequestContext, question: models.Question
-) -> schemas.Question:
-    question_data = questions_service.question_schema(ctx, question)
-    if question_data is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The question doesn't exist in the system.",
-        )
-    return question_data
 
 
 @router.get(
@@ -62,23 +30,7 @@ def get_question(
     """
     Get question in one of current_user's belonging sites.
     """
-    question = questions_service.get_readable_question(
-        ctx.get_db(),
-        uuid=uuid,
-        principal_id=ctx.principal_id,
-        ctx=ctx,
-    )
-    if question is None:
-        raise HTTPException_(
-            status_code=404,
-            detail="The question doesn't exist in the system.",
-        )
-    if question.is_hidden:
-        raise HTTPException_(
-                status_code=403,
-                detail="Not allowed to access this quesion",
-        )
-    return _get_question_data(ctx, question)
+    return questions_service.get_question(ctx, uuid=uuid)
 
 
 @router.post("/{uuid}/views/", response_model=schemas.GenericResponse)
@@ -88,8 +40,7 @@ def bump_views_counter(
     ctx: RequestContext = Depends(deps.get_request_context),
     _current_user_id: Optional[int] = Depends(deps.try_get_current_user_id),
 ) -> Any:
-    question = questions_service.get_question_model_http(ctx.get_db(), uuid)
-    view_counters.add_view_async(ctx, "question", question.id)
+    questions_service.bump_views(ctx, uuid=uuid)
     return schemas.GenericResponse()
 
 
@@ -106,18 +57,9 @@ def get_question_answers(
     """
     Get question's answers' previews.
     """
-    question = questions_service.get_question_model_http(ctx.get_db(), uuid)
-    if not user_in_site(
-        ctx.get_db(),
-        site=question.site,
-        user_id=current_user_id,
-        op_type=OperationType.ReadSite,
-    ):
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    return _get_answers(ctx, question)
+    return questions_service.get_question_answers(
+        ctx, uuid=uuid, principal_id=current_user_id
+    )
 
 
 @router.post("/", response_model=schemas.Question)
@@ -131,37 +73,11 @@ def create_question(
     """
     Create new question authored by the current user in one of the belonging sites.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    crud.audit_log.create_with_user(
-        db,
-        ipaddr=client_ip(request),
-        user_id=current_user.id,
-        api="post question",
-        request_info={"question_in": jsonable_encoder(question_in)},
-    )
-
-    site = get_site(db, question_in.site_uuid)
-    check_user_in_site(
-        db,
-        site=site,
-        user_id=current_user.id,
-        op_type=OperationType.WriteSiteQuestion,
-    )
-    if current_user.remaining_coins < site.create_question_coin_deduction:
-        raise HTTPException_(
-            status_code=400,
-            detail="Insufficient coins.",
-        )
-    new_question = crud.question.create_with_author(
-        db, obj_in=question_in, author_id=current_user.id
-    )
-    # Subscribed to the authored question automatically
-    current_user = crud.user.subscribe_question(
-        db, db_obj=current_user, question=new_question
+    new_question, data = questions_service.create_question(
+        ctx, question_in=question_in, ipaddr=client_ip(request)
     )
     background_tasks.add_task(postprocess_new_question, new_question.id)
-    return questions_service.question_schema(ctx, new_question)
+    return data
 
 
 @router.put("/{uuid}", response_model=schemas.Question)
@@ -177,75 +93,15 @@ def update_question(
     """
     Update question in one of current_user's belonging sites as member.
     """
-    db = ctx.get_db()
-    crud.audit_log.create_with_user(
-        db,
+    new_question, data = questions_service.update_question(
+        ctx,
+        uuid=uuid,
+        question_in=question_in,
+        current_user_id=current_user_id,
         ipaddr=client_ip(request),
-        user_id=current_user_id,
-        api="put question",
-        request_info={"question_in": jsonable_encoder(question_in), "uuid": uuid},
     )
-
-    question = crud.question.get_by_uuid(db, uuid=uuid)
-    if question is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The question doesn't exist in the system.",
-        )
-    if current_user_id != question.author_id:
-        if not user_in_site(
-            db,
-            user_id=current_user_id,
-            site=question.site,
-            op_type=OperationType.WriteSiteQuestion,
-        ):
-            raise HTTPException_(
-                status_code=400,
-                detail="Unauthorized.",
-            )
-    editor_id = question.editor_id
-    if editor_id is None:
-        editor_id = question.author_id
-    archive = models.QuestionArchive(
-        question_id=question.id,
-        title=question.title,
-        description=question.description,
-        description_text=question.description_text,
-        description_editor=question.description_editor,
-        created_at=question.updated_at,
-        editor_id=editor_id,
-    )
-    archive.topics = question.topics
-    db.add(archive)
-    question.archives.append(archive)
-    db.commit()
-    if question_in.topic_uuids is not None:
-        new_topics = []
-        for topic_uuid in question_in.topic_uuids:
-            topic = crud.topic.get_by_uuid(db, uuid=topic_uuid)
-            if topic is None:
-                raise HTTPException_(
-                    status_code=400,
-                    detail="The topic doesn't exist.",
-                )
-            new_topics.append(topic)
-        question_in.topic_uuids = None
-        question = crud.question.update_topics(
-            db, db_obj=question, new_topics=new_topics
-        )
-    question_in_dict = question_in.dict(exclude_none=True)
-    question_in_dict["editor_id"] = current_user_id
-    question_in_dict["updated_at"] = datetime.datetime.now(tz=datetime.timezone.utc)
-    if question_in.desc:
-        question_in_dict["description"] = question_in.desc.source
-        question_in_dict["description_editor"] = question_in.desc.editor
-        question_in_dict["description_text"] = question_in.desc.rendered_text
-    else:
-        question_in_dict["description"] = None
-        question_in_dict["description_text"] = None
-    new_question = crud.question.update(db, db_obj=question, obj_in=question_in_dict)
     background_tasks.add_task(postprocess_updated_question, new_question.id)
-    return questions_service.question_schema(ctx, new_question)
+    return data
 
 
 @router.get("/{uuid}/archives/", response_model=List[schemas.QuestionArchive])
@@ -272,19 +128,7 @@ def hide_question(
     *,
     uuid: str,
 ) -> Any:
-    question = questions_service.get_question_model_http(ctx.get_db(), uuid)
-    if (
-        question.site.moderator_id != ctx.principal_id
-        and question.author_id != ctx.principal_id
-    ):
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    question = crud.question.update(
-        ctx.get_db(), db_obj=question, obj_in={"is_hidden": True}
-    )
-    return questions_service.question_schema(ctx, question)
+    return questions_service.hide_question(ctx, uuid=uuid)
 
 
 @router.post(
@@ -296,43 +140,7 @@ def invite_answer(
     uuid: str,
     user_uuid: str,
 ) -> Any:
-    question = questions_service.get_question_model_http(ctx.get_db(), uuid)
-    check_user_in_site(
-        ctx.get_db(),
-        site=question.site,
-        user_id=ctx.unwrapped_principal_id(),
-        op_type=OperationType.ReadSite,
-    )
-    invited_user = crud.user.get_by_uuid(ctx.get_db(), uuid=user_uuid)
-    if invited_user is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The user doesn't exist in the system.",
-        )
-    if invited_user.id == ctx.unwrapped_principal_id():
-        raise HTTPException_(
-            status_code=400,
-            detail="You can't invite yourself.",
-        )
-    check_user_in_site(
-        ctx.get_db(),
-        site=question.site,
-        user_id=invited_user.id,
-        op_type=OperationType.WriteSiteQuestion,
-    )
-    utc_now = datetime.datetime.now(tz=datetime.timezone.utc)
-    crud.notification.create_with_content(
-        ctx.broker,
-        receiver_id=invited_user.id,
-        event=EventInternal(
-            created_at=utc_now,
-            content=InviteAnswerInternal(
-                subject_id=ctx.unwrapped_principal_id(),
-                question_id=question.id,
-                user_id=invited_user.id,
-            ),
-        ),
-    )
+    questions_service.invite_answer(ctx, uuid=uuid, user_uuid=user_uuid)
     return schemas.GenericResponse()
 
 
@@ -345,68 +153,7 @@ def upvote_question(
     """
     Upvote question as the current user.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    question = questions_service.get_question_model_http(ctx.get_db(), uuid)
-    check_user_in_site(
-        db,
-        site=question.site,
-        user_id=current_user.id,
-        op_type=OperationType.ReadSite,
-    )
-    upvoted = (
-        db.query(models.QuestionUpvotes)
-        .filter_by(question_id=question.id, voter_id=current_user.id, cancelled=False)
-        .first()
-        is not None
-    )
-    if not upvoted:
-        if current_user.id == question.author_id:
-            raise HTTPException_(
-                status_code=400,
-                detail="Author can't upvote authored question.",
-            )
-        if current_user.remaining_coins < question.site.upvote_question_coin_deduction:
-            raise HTTPException_(
-                status_code=400,
-                detail="Insufficient coins.",
-            )
-        upvoted_before = (
-            db.query(models.QuestionUpvotes)
-            .filter_by(question_id=question.id, voter_id=current_user.id)
-            .first()
-            is not None
-        )
-        # Don't swap the statements before and after!
-        question = crud.question.upvote(db, db_obj=question, voter=current_user)
-        utc_now = datetime.datetime.now(tz=datetime.timezone.utc)
-        if not upvoted_before:
-            crud.coin_payment.make_payment(
-                db,
-                obj_in=schemas.CoinPaymentCreate(
-                    payee_id=question.author_id,
-                    amount=question.site.upvote_question_coin_deduction,
-                    event_json=EventInternal(
-                        created_at=utc_now,
-                        content=UpvoteQuestionInternal(
-                            subject_id=current_user.id,
-                            question_id=question.id,
-                        ),
-                    ).json(),
-                ),
-                payer=current_user,
-                payee=question.author,
-            )
-        db.commit()
-        db.refresh(question)
-    valid_upvotes = (
-        db.query(models.QuestionUpvotes)
-        .filter_by(question_id=question.id, cancelled=False)
-        .count()
-    )
-    return schemas.QuestionUpvotes(
-        question_uuid=question.uuid, count=valid_upvotes, upvoted=True
-    )
+    return questions_service.upvote_question(ctx, uuid=uuid)
 
 
 @router.delete("/{uuid}/upvotes/", response_model=schemas.QuestionUpvotes)
@@ -418,38 +165,7 @@ def cancel_upvote_question(
     """
     Cancel upvote for question as the current user.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    question = crud.question.get_by_uuid(db, uuid=uuid)
-    if question is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The question doesn't exist in the system.",
-        )
-    check_user_in_site(
-        db,
-        site=question.site,
-        user_id=current_user.id,
-        op_type=OperationType.ReadSite,
-    )
-    upvoted = (
-        db.query(models.QuestionUpvotes)
-        .filter_by(question_id=question.id, voter_id=current_user.id, cancelled=False)
-        .first()
-        is not None
-    )
-    if upvoted:
-        question = crud.question.cancel_upvote(db, db_obj=question, voter=current_user)
-        db.commit()
-        db.refresh(question)
-    valid_upvotes = (
-        db.query(models.QuestionUpvotes)
-        .filter_by(question_id=question.id, cancelled=False)
-        .count()
-    )
-    return schemas.QuestionUpvotes(
-        question_uuid=question.uuid, count=valid_upvotes, upvoted=False
-    )
+    return questions_service.cancel_upvote_question(ctx, uuid=uuid)
 
 
 @router.get("/{uuid}/page", response_model=schemas.QuestionPage)
@@ -461,81 +177,4 @@ def get_question_page(
     ctx: RequestContext = Depends(deps.get_request_context),
     uuid: str,
 ) -> Any:
-    current_user_id = ctx.principal_id
-    question = questions_service.get_readable_question(
-        ctx.get_db(),
-        uuid=uuid,
-        principal_id=current_user_id,
-        ctx=ctx,
-    )
-    if question is None:
-        from chafan_core.app.services import audit as audit_service
-
-        audit_service.create_audit(
-            ctx.get_db(),
-            api=f"get_question_page {uuid} retrieved None",
-            request=request,
-            user_id=current_user_id,
-        )
-        raise HTTPException_(
-                status_code=404,
-                detail="No such question"
-            )
-    question_data = _get_question_data(ctx, question)
-    flags = schemas.QuestionPageFlags()
-    if ctx.principal_id:
-        if question.author_id == ctx.principal_id:
-            flags.editable = True
-            flags.hideable = True
-        if user_in_site(
-            ctx.get_db(),
-            site=question.site,
-            user_id=ctx.principal_id,
-            op_type=OperationType.WriteSiteQuestion,
-        ):
-            flags.editable = True
-        if user_in_site(
-            ctx.get_db(),
-            site=question.site,
-            user_id=ctx.principal_id,
-            op_type=OperationType.WriteSiteAnswer,
-        ):
-            flags.answer_writable = True
-        if user_in_site(
-            ctx.get_db(),
-            site=question.site,
-            user_id=ctx.principal_id,
-            op_type=OperationType.WriteSiteComment,
-        ):
-            flags.comment_writable = True
-        if question.site.moderator_id == ctx.principal_id:
-            flags.is_mod = True
-            flags.hideable = True
-    # TODO 2025-07-08 This is hacky. The whole logic of question flags needs to be reviewed and simplified.
-    if question.site.public_writable_answer:
-        flags.answer_writable = True
-    return schemas.QuestionPage(
-        question=question_data,
-        # TODO: continuation
-        # TODO: rethink the internal caching
-        full_answers=filter_not_none(
-            [
-                answers_service.answer_schema(ctx, answer)
-                for answer in rank_answers(
-                    question.answers, principal_id=ctx.principal_id
-                )
-            ]
-        ),
-        answer_previews=filter_not_none(
-            [
-                ctx.preview_of_answer(answer)
-                for answer in rank_answers(
-                    question.answers, principal_id=ctx.principal_id
-                )
-            ]
-        ),
-        question_subscription=questions_service.get_question_subscription(
-            ctx, question
-        ),
-        flags=flags,
-    )
+    return questions_service.get_question_page(ctx, uuid=uuid, request=request)
