+++ chafan_core/app/api/api_v1/endpoints/submissions.py
@@ -1,28 +1,17 @@
-import datetime
-from typing import Any, List, Optional, Union
+from typing import Any, List, Optional
 
-from fastapi import APIRouter, Depends, Request, BackgroundTasks
-from fastapi.encoders import jsonable_encoder
+from fastapi import APIRouter, BackgroundTasks, Depends, Request
 from sqlalchemy.orm import Session
 
-import logging
-logger = logging.getLogger(__name__)
-
-
-from chafan_core.app import crud, models, schemas, view_counters
+from chafan_core.app import schemas
 from chafan_core.app.api import deps
-from chafan_core.app.services import submissions as submissions_service
+from chafan_core.app.common import client_ip
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.common import OperationType, client_ip
-from chafan_core.app.endpoint_utils import get_site
-from chafan_core.app.responders.archives import submission_archive_schema_from_orm
-from chafan_core.app.user_permission import check_user_in_site
-from chafan_core.app.schemas.event import EventInternal, UpvoteSubmissionInternal
+from chafan_core.app.services import submissions as submissions_service
 from chafan_core.app.services.postprocess import (
     postprocess_new_submission,
     postprocess_updated_submission,
 )
-from chafan_core.utils.base import HTTPException_, filter_not_none
 
 router = APIRouter()
 
@@ -49,24 +38,7 @@ def get_submission(
     """
     Get submission in one of current_user's belonging sites.
     """
-    logger.info("get submission " + uuid)
-    submission = crud.submission.get_by_uuid(ctx.get_db(), uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    submission_data: Optional[
-        schemas.Submission
-    ] = None
-    # TODO didn't check principal id
-    submission_data = submissions_service.submission_schema(ctx, submission)
-    if submission_data is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    return submission_data
+    return submissions_service.get_submission(ctx, uuid=uuid)
 
 
 @router.get("/{uuid}/upvotes/", response_model=schemas.SubmissionUpvotes)
@@ -76,26 +48,8 @@ def get_submission_upvotes(
     uuid: str,
     current_user_id: Optional[int] = Depends(deps.try_get_current_user_id),
 ) -> Any:
-    submission = crud.submission.get_by_uuid(db, uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    valid_upvotes = crud.submission.count_upvotes(db, submission)
-    if current_user_id:
-        upvoted = (
-            db.query(models.SubmissionUpvotes)
-            .filter_by(
-                submission_id=submission.id, voter_id=current_user_id, cancelled=False
-            )
-            .first()
-            is not None
-        )
-    else:
-        upvoted = False
-    return schemas.SubmissionUpvotes(
-        submission_uuid=submission.uuid, count=valid_upvotes, upvoted=upvoted
+    return submissions_service.get_upvotes(
+        db, uuid=uuid, principal_id=current_user_id
     )
 
 
@@ -105,43 +59,8 @@ def bump_views_counter(
     uuid: str,
     ctx: RequestContext = Depends(deps.get_request_context),
 ) -> Any:
-    submission = crud.submission.get_by_uuid(ctx.get_db(), uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    view_counters.add_view_async(ctx, "submission", submission.id)
+    submissions_service.bump_views(ctx, uuid=uuid)
     return schemas.GenericResponse()
-
-
-def _create_submission(
-    ctx: RequestContext,
-    submission_in: schemas.SubmissionCreate,
-    author: models.User,
-    background_tasks: BackgroundTasks,
-) -> schemas.Submission:
-    db = ctx.get_db()
-    site = get_site(db, submission_in.site_uuid)
-    check_user_in_site(
-        db,
-        site=site,
-        user_id=author.id,
-        op_type=OperationType.WriteSiteSubmission,
-    )
-    if author.remaining_coins < site.create_submission_coin_deduction:
-        raise HTTPException_(
-            status_code=400,
-            detail="Insufficient coins.",
-        )
-
-    new_submission = crud.submission.create_with_author(
-        db, obj_in=submission_in, author_id=author.id
-    )
-    background_tasks.add_task(postprocess_new_submission, new_submission.id)
-    data = submissions_service.submission_schema(ctx, new_submission)
-    assert data is not None
-    return data
 
 
 @router.post("/", response_model=schemas.Submission)
@@ -156,64 +75,14 @@ def create_submission(
     Create new submission authored by the current user in one of the belonging sites.
     """
     current_user = ctx.get_current_active_user()
-    crud.audit_log.create_with_user(
-        ctx.get_db(),
+    new_submission, data = submissions_service.create_submission(
+        ctx,
+        submission_in=submission_in,
+        author=current_user,
         ipaddr=client_ip(request),
-        user_id=current_user.id,
-        api="post submission",
-        request_info={"submission_in": jsonable_encoder(submission_in)},
     )
-    return _create_submission(
-        ctx, submission_in, current_user, background_tasks=background_tasks
-    )
-
-
-def _update_submission(
-    ctx: RequestContext,
-    *,
-    submission: models.Submission,
-    submission_in: schemas.SubmissionUpdate,
-    background_tasks: BackgroundTasks,
-) -> Optional[schemas.Submission]:
-    db = ctx.get_db()
-    archive = models.SubmissionArchive(
-        submission_id=submission.id,
-        title=submission.title,
-        description=submission.description,
-        description_editor=submission.description_editor,
-        description_text=submission.description_text,
-        topic_uuids=[t.uuid for t in submission.topics],
-        created_at=submission.updated_at,
-    )
-    db.add(archive)
-    submission.archives.append(archive)
-    db.commit()
-    if submission_in.topic_uuids is not None:
-        new_topics = []
-        for topic_uuid in submission_in.topic_uuids:
-            topic = crud.topic.get_by_uuid(db, uuid=topic_uuid)
-            if topic is None:
-                raise HTTPException_(
-                    status_code=400,
-                    detail="The topic doesn't exist.",
-                )
-            new_topics.append(topic)
-        submission_in.topic_uuids = None
-        submission = crud.submission.update_topics(
-            db, db_obj=submission, new_topics=new_topics
-        )
-    submission_in_dict = submission_in.dict(exclude_unset=True)
-    submission_in_dict["updated_at"] = datetime.datetime.now(tz=datetime.timezone.utc)
-    if submission_in.desc:
-        del submission_in_dict["desc"]
-        submission_in_dict["description"] = submission_in.desc.source
-        submission_in_dict["description_editor"] = submission_in.desc.editor
-        submission_in_dict["description_text"] = submission_in.desc.rendered_text
-    new_submission = crud.submission.update(
-        db, db_obj=submission, obj_in=submission_in_dict
-    )
-    background_tasks.add_task(postprocess_updated_submission, new_submission.id)
-    return submissions_service.submission_schema(ctx, new_submission)
+    background_tasks.add_task(postprocess_new_submission, new_submission.id)
+    return data
 
 
 @router.put("/{uuid}", response_model=schemas.Submission)
@@ -228,31 +97,14 @@ def update_submission(
     """
     Update submission as author.
     """
-    crud.audit_log.create_with_user(
-        ctx.get_db(),
-        ipaddr=client_ip(request),
-        user_id=ctx.unwrapped_principal_id(),
-        api="post submission",
-        request_info={"submission_in": jsonable_encoder(submission_in), "uuid": uuid},
-    )
-
-    submission = crud.submission.get_by_uuid(ctx.get_db(), uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    if ctx.unwrapped_principal_id() != submission.author_id:
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    return _update_submission(
+    new_submission, data = submissions_service.update_submission(
         ctx,
-        submission=submission,
+        uuid=uuid,
         submission_in=submission_in,
-        background_tasks=background_tasks,
+        ipaddr=client_ip(request),
     )
+    background_tasks.add_task(postprocess_updated_submission, new_submission.id)
+    return data
 
 
 @router.get("/{uuid}/archives/", response_model=List[schemas.SubmissionArchive])
@@ -265,19 +117,9 @@ def get_submission_archives(
     """
     Get answer's archives as its author.
     """
-    submission = crud.submission.get_by_uuid(db, uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    check_user_in_site(
-        db,
-        site=submission.site,
-        user_id=current_user_id,
-        op_type=OperationType.ReadSite,
+    return submissions_service.list_archives(
+        db, uuid=uuid, principal_id=current_user_id
     )
-    return [submission_archive_schema_from_orm(a) for a in submission.archives]
 
 
 @router.get("/{uuid}/suggestions/", response_model=List[schemas.SubmissionSuggestion])
@@ -293,27 +135,9 @@ def get_submission_suggestions(
 def hide_submission(
     ctx: RequestContext = Depends(deps.get_request_context_logged_in),
     *,
-    db: Session = Depends(deps.get_db),
     uuid: str,
 ) -> Any:
-    submission = crud.submission.get_by_uuid(db, uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    if (
-        submission.site.moderator_id != ctx.principal_id
-        and submission.author_id != ctx.principal_id
-    ):
-        raise HTTPException_(
-            status_code=400,
-            detail="Unauthorized.",
-        )
-    submission = crud.submission.update(
-        db, db_obj=submission, obj_in={"is_hidden": True}
-    )
-    return submissions_service.submission_schema(ctx, submission)
+    return submissions_service.hide_submission(ctx, uuid=uuid)
 
 
 @router.post("/{uuid}/upvotes/", response_model=schemas.SubmissionUpvotes)
@@ -325,78 +149,7 @@ def upvote_submission(
     """
     Upvote submission as the current user.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    submission = crud.submission.get_by_uuid(db, uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    check_user_in_site(
-        db,
-        site=submission.site,
-        user_id=current_user.id,
-        op_type=OperationType.ReadSite,
-    )
-    upvoted = (
-        db.query(models.SubmissionUpvotes)
-        .filter_by(
-            submission_id=submission.id, voter_id=current_user.id, cancelled=False
-        )
-        .first()
-        is not None
-    )
-    if not upvoted:
-        if current_user.id == submission.author_id:
-            raise HTTPException_(
-                status_code=400,
-                detail="Author can't upvote authored submission.",
-            )
-        if (
-            current_user.remaining_coins
-            < submission.site.upvote_submission_coin_deduction
-        ):
-            raise HTTPException_(
-                status_code=400,
-                detail="Insufficient coins.",
-            )
-        upvoted_before = (
-            db.query(models.SubmissionUpvotes)
-            .filter_by(submission_id=submission.id, voter_id=current_user.id)
-            .first()
-            is not None
-        )
-        # Don't swap the statements before and after!
-        submission = crud.submission.upvote(db, db_obj=submission, voter=current_user)
-        utc_now = datetime.datetime.now(tz=datetime.timezone.utc)
-        if not upvoted_before:
-            crud.coin_payment.make_payment(
-                db,
-                obj_in=schemas.CoinPaymentCreate(
-                    payee_id=submission.author_id,
-                    amount=submission.site.upvote_submission_coin_deduction,
-                    event_json=EventInternal(
-                        created_at=utc_now,
-                        content=UpvoteSubmissionInternal(
-                            subject_id=current_user.id,
-                            submission_id=submission.id,
-                        ),
-                    ).json(),
-                ),
-                payer=current_user,
-                payee=submission.author,
-            )
-        db.commit()
-        db.refresh(submission)
-    valid_upvotes = (
-        db.query(models.SubmissionUpvotes)
-        .filter_by(submission_id=submission.id, cancelled=False)
-        .count()
-    )
-    return schemas.SubmissionUpvotes(
-        submission_uuid=submission.uuid, count=valid_upvotes, upvoted=True
-    )
+    return submissions_service.upvote_submission(ctx, uuid=uuid)
 
 
 @router.delete("/{uuid}/upvotes/", response_model=schemas.SubmissionUpvotes)
@@ -408,39 +161,4 @@ def cancel_upvote_submission(
     """
     Cancel upvote for submission as the current user.
     """
-    current_user = ctx.get_current_active_user()
-    db = ctx.get_db()
-    submission = crud.submission.get_by_uuid(db, uuid=uuid)
-    if submission is None:
-        raise HTTPException_(
-            status_code=400,
-            detail="The submission doesn't exist in the system.",
-        )
-    check_user_in_site(
-        db,
-        site=submission.site,
-        user_id=current_user.id,
-        op_type=OperationType.ReadSite,
-    )
-    upvoted = (
-        db.query(models.SubmissionUpvotes)
-        .filter_by(
-            submission_id=submission.id, voter_id=current_user.id, cancelled=False
-        )
-        .first()
-        is not None
-    )
-    if upvoted:
-        submission = crud.submission.cancel_upvote(
-            db, db_obj=submission, voter=current_user
-        )
-        db.commit()
-        db.refresh(submission)
-    valid_upvotes = (
-        db.query(models.SubmissionUpvotes)
-        .filter_by(submission_id=submission.id, cancelled=False)
-        .count()
-    )
-    return schemas.SubmissionUpvotes(
-        submission_uuid=submission.uuid, count=valid_upvotes, upvoted=False
-    )
+    return submissions_service.cancel_upvote_submission(ctx, uuid=uuid)
